requirements.txt
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
bm25s>=0.1.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
lxml>=4.9.0
accelerate>=0.25.0
